requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Caching==2.3.0
Flask-WTF==1.2.2
Flask-Limiter==3.7.0
flask-talisman==1.1.0
gunicorn==22.0.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
minio==7.2.10
Pillow==10.4.0
redis>=5.0.0
black==24.4.2
ruff==0.4.4